bi-directional streaming for normal Queries
