+++ grpc-proto/proto/stargate.proto
@@ -27,7 +27,13 @@ service Stargate {
   // Executes a single CQL query.
   rpc ExecuteQuery(Query) returns (Response) {}
 
+  // Executes a bi-directional streaming for CQL queries.
+  rpc ExecuteQueryStream(stream Query) returns (stream Response) {}
+
   // Executes a batch of CQL queries.
   rpc ExecuteBatch(Batch) returns (Response) {}
+
+  // Executes a bi-directional streaming for batches of CQL queries.
+  rpc ExecuteBatchStream(stream Batch) returns (stream Response) {}
 }
